+++ frontend/components/feedback-task/container/questions/form/shared-components/label-selection/LabelSelection.component.vue
@@ -56,7 +56,8 @@
           @keydown.tab="expandLabelsOnTab(index)"
         />
         <BaseTooltip
-          :text="isSuggested(option) ? $t('suggestion.name') : null"
+          :title="isSuggested(option) ? $t('suggestion.name') : ''"
+          :text="getSuggestedAgent(option)"
           minimalist
         >
           <label
@@ -70,13 +71,17 @@
             :title="option.text"
           >
             <span class="key" v-text="keyboards[option.id]" />
-            <span>{{ option.text }}</span>
-            <svgicon
-              v-if="isSuggested(option)"
-              class="label-text__suggestion-icon"
-              name="suggestion"
-            /> </label
-        ></BaseTooltip>
+            <span class="label-text__text">{{ option.text }}</span>
+            <span v-if="isSuggested(option)" class="label-text__suggestion">
+              <svgicon class="label-text__suggestion__icon" name="suggestion" />
+              <span
+                v-if="getSuggestedScore(option)"
+                class="label-text__suggestion__score"
+                v-text="getSuggestedScore(option)"
+              />
+            </span>
+          </label>
+        </BaseTooltip>
       </div>
     </transition-group>
     <i class="no-result" v-if="!filteredOptions.length" />
@@ -113,6 +118,10 @@ export default {
       required: true,
     },
     multiple: {
+      type: Boolean,
+      default: () => false,
+    },
+    suggestionFirst: {
       type: Boolean,
       default: () => false,
     },
@@ -178,11 +187,40 @@ export default {
         .filter((option) => option.isSelected);
     },
     visibleOptions() {
-      if (this.isExpanded) return this.filteredOptions;
+      if (!this.suggestionFirst) {
+        if (this.isExpanded) return this.filteredOptions;
 
-      return this.filteredOptions
-        .slice(0, this.maxOptionsToShowBeforeCollapse)
-        .concat(this.remainingVisibleOptions);
+        return this.filteredOptions
+          .slice(0, this.maxOptionsToShowBeforeCollapse)
+          .concat(this.remainingVisibleOptions);
+      }
+
+      const suggestedOptions = this.filteredOptions
+        .filter((v) => this.suggestion && this.suggestion.isSuggested(v.value))
+        .sort((a, b) => {
+          const isASuggested = this.suggestion.getSuggestion(a.value);
+          const isBSuggested = this.suggestion.getSuggestion(b.value);
+
+          return isASuggested?.score > isBSuggested?.score ? -1 : 1;
+        });
+
+      const noSuggestedOptions = this.filteredOptions.filter(
+        (v) => !this.suggestion || !this.suggestion.isSuggested(v.value)
+      );
+
+      if (this.isExpanded) {
+        return [...suggestedOptions, ...noSuggestedOptions];
+      }
+
+      const options = [
+        ...suggestedOptions.filter((o) => o.isSelected),
+        ...noSuggestedOptions.filter((o) => o.isSelected),
+      ];
+
+      return options.slice(
+        0,
+        Math.max(options.length, this.maxOptionsToShowBeforeCollapse)
+      );
     },
     numberToShowInTheCollapseButton() {
       return this.filteredOptions.length - this.visibleOptions.length;
@@ -309,6 +347,12 @@ export default {
     isSuggested(option) {
       return this.suggestion?.isSuggested(option.value);
     },
+    getSuggestedScore(option) {
+      return this.suggestion?.getSuggestion(option.value)?.score?.fixed;
+    },
+    getSuggestedAgent(option) {
+      return this.suggestion?.getSuggestion(option.value)?.agent;
+    },
   },
   setup(props) {
     return useLabelSelectionViewModel(props);
@@ -391,20 +435,27 @@ $label-dark-color: palette(purple, 200);
   transition: all 0.2s ease-in-out;
   cursor: pointer;
   user-select: none;
-
-  &__suggestion-icon {
-    flex-shrink: 0;
-    width: 10px;
-    height: 10px;
-  }
-
-  span {
+  &__text {
     white-space: nowrap;
     overflow: hidden;
     text-overflow: ellipsis;
     min-width: 0;
     &:hover {
       direction: rtl;
+    }
+  }
+  &__suggestion {
+    display: flex;
+    flex-shrink: 0;
+    align-items: center;
+    gap: calc($base-space / 2);
+    &__score {
+      @include font-size(11px);
+    }
+    &__icon {
+      flex-shrink: 0;
+      width: 10px;
+      height: 10px;
     }
   }