+++ frontend/pages/dataset/_id/annotation-mode/index.component.spec.js
@@ -35,6 +35,7 @@ describe("AnnotationModePage", () => {
         "BaseModal",
         "DatasetTrainComponent",
         "BaseLoading",
+        "PersistentStorageBanner",
       ],
       mocks: {
         $route,
@@ -68,6 +69,7 @@ describe("AnnotationModePage", () => {
         "BaseModal",
         "DatasetTrainComponent",
         "BaseLoading",
+        "PersistentStorageBanner",
       ],
       mocks: {
         $route,