+++ frontend/translation/en.js
@@ -23,6 +23,7 @@ export default {
   description: "Description",
   labels: "Labels",
   useMarkdown: "Use Markdown",
+  suggestionFirst: "Suggestion first",
   visibleForAnnotators: "Visible for annotators",
   allowExtraMetadata: "Allow extra metadata",
   extraMetadata: "Extra metadata",
@@ -130,6 +131,7 @@ export default {
     discard: "Discard",
     submit: "Submit",
     draft: "Save as draft",
+    write: "Write",
   },
   sorting: {
     addOtherField: "+ Add another field",
@@ -151,7 +153,6 @@ export default {
       agent: "Agent",
     },
     plural: "Suggestions",
-    "suggested-rank": `Suggested rank`,
     name: `Suggestion`,
   },
   similarity: {